Refactor forest_client_integration.py file directory path
