+++ server/backend/api/app/routers/router_forest_client.py
@@ -6,7 +6,7 @@ from api.app.routers.router_utils import get_api_instance_env
 from api.app.schemas import FamForestClientSchema
 from fastapi import APIRouter, Depends, Query
 
-from server.backend.api.app.integration.forest_client.forest_client_integration import \
+from server.backend.api.app.integration.forest_client_integration import \
     ForestClientIntegrationService
 
 LOGGER = logging.getLogger(__name__)
